feat: Jest test wrapper component created, CypressTestUtils folder relocated
