+++ packages/utils/src/index.ts
@@ -15,9 +15,10 @@ export * from './Registry';
 export { default as routerParams } from './RouterParams';
 export { default as RowLoader } from './RowLoader';
 export { useInventory } from './useInventory';
-export * from './CypressUtils';
+export * from './TestingUtils/CypressUtils';
 export * from './useInsightsNavigate';
 export * from './useExportPDF';
 export * from './usePromiseQueue';
 export * from './useFetchBatched';
 export * from './useSuspenseLoader';
+export * from './TestingUtils/JestUtils';
